Add end note plug
I created a plugin that creates end notes and a reference list. Aside from appending the reference list below the footer it works as expected. The feature is "binned" because it caused the build times to increase x20 for a single blog post with just 20 links. I'll investigate in the future if this feature is necessary.

It also printed some css to the console which is also unresolved.